Stop running lint on all files when no changes
Stops our lint from running on all of the files in `salt/` or `test/`
when no files were modified. By default, when tox is told to run with no
positional arguments, it'll run on all files we'd look at. When run with
positional arguments it'll run only on those files. This makes it so
that tox, by default, will only run lint checks on commits that actually
change the files in `salt/` or `test/`.