Update documentation for #48694 and add deprecation warning
There was a behavior change when using `salt-run cache.grains` beginning
with the 2017.7.5 release.

The behavior change happened in PR #45588, which was back-ported to the
2017.7 branch in PR #45751. This was done in order to fix the bug reported
in #45489.

In order to be helpful to users, a deprecation warning was added. If ``tgt``
is not passed, no grains will be listed. Instead of suddenly requiring a ``tgt``
argument in a dot release, let's keep this optional arg around until Sodium.

When Salt Sodium is released, we can remove the warning and change ``tgt`` to be
a mandatory argument.

Fixes the last parts of #48694.