Increase centos7-py2 kitchen pr timeout to 8 hours
The centos7-py2 tests have been consistently failing to finish before
the timeout. I am increasing the timeout for these tests to 8 hours in
order to see if we can finish a few of them. They should get much closer
to finishing as the branch tests have been finishing at around the 6:30
or 7:00 mark sometimes.