Don't import obsolete Directive class from Sphinx
Directive was apparently removed from Sphinx and made part of docutils.
For some reason, in this saltdomain extension, we were importing both
from docutils and sphinx. This worked well enough while Directive was
still only deprecated, but it appears to recently have been removed,
meaning that it now just causes an ImportError.

This commit removes the redundant import.