Updates to states and modules for docker
Due to changes in docker modules in `salt v2017.7`, the module and state files named `dockerng` are not longer available but have been split into several different files.  The custom changes we originally added to salt-formulas to override the default modules are now being ported over and applied to the new file structure.   As a result, `states/_modules/dockerng.py` and `states/_states/dockerng.py` in salt-formulas will be removed.